feat(ka-routes)!: Stage 4 — delete legacy assertion.ts; preserve /api/kc reads
Removes packages/cli/src/daemon/routes/assertion.ts (3909 lines) and its dispatch
in handle-request.ts. The /api/assertion/* HTTP surface is gone — everything now
flows through /api/knowledge-assets/*.

Two read-only chain-metadata routes that happened to live in assertion.ts but are
NOT part of the assertion lifecycle (GET /api/kc/:id, GET /api/kc/:id/author) are
extracted verbatim into the new routes/kc-chain-metadata.ts module and dispatched
separately, preserving their exact paths/shapes. kc-author-route.e2e.test.ts
repointed to the new module.

BREAKING: /api/assertion/* routes no longer exist. All first-party clients,
adapters, scripts, docs, and tests were migrated to /api/knowledge-assets/* in
Stages 2-3.
