+++ packages/cli/src/daemon/handle-request.ts
@@ -321,8 +321,8 @@ import { handleHermesRoutes } from './routes/hermes.js';
 import { handleMemoryRoutes } from './routes/memory.js';
 import { handlePublisherRoutes } from './routes/publisher.js';
 import { handleContextGraphRoutes } from './routes/context-graph.js';
-import { handleAssertionRoutes } from './routes/assertion.js';
 import { handleKnowledgeAssetsRoutes } from './routes/knowledge-assets.js';
+import { handleKcChainMetadataRoutes } from './routes/kc-chain-metadata.js';
 import { handleQueryRoutes } from './routes/query.js';
 import { handleLocalAgentsRoutes } from './routes/local-agents.js';
 import { handleEpcisRoutes } from './routes/epcis.js';
@@ -433,7 +433,7 @@ export async function handleRequest(
   await handleKnowledgeAssetsRoutes(ctx);
   if (res.writableEnded) return;
 
-  await handleAssertionRoutes(ctx);
+  await handleKcChainMetadataRoutes(ctx);
   if (res.writableEnded) return;
 
   await handleQueryRoutes(ctx);
